Don't fail when POSH_THEMES_PATH is null

+++ install.ps1
@@ -9,7 +9,7 @@ if ( !( Test-Path $profileDir -PathType Container ) ) {
 Copy-Item .\ps\profile\profile.ps1 "${profileDir}\Microsoft.PowerShell_profile.ps1"
 Copy-Item .\ps\profile\profile.ps1 "${profileDir}\Microsoft.VSCode_profile.ps1"
 
-if ( ( Test-Path $Env:POSH_THEMES_PATH -PathType Container ) ) {
+if ( $Env:POSH_THEMES_PATH -and ( Test-Path $Env:POSH_THEMES_PATH -PathType Container ) ) {
     Copy-Item -Force ./ps/oh-my-posh/git-on-up.omp.json $Env:POSH_THEMES_PATH
 }
